Rename bot settings section to general settings
