+++ src/views/generalSettings/components/CategoryList/index.css
