+++ src/views/generalSettings/components/CategoryList/index.js
