+++ src/views/generalSettings/components/NewCategoryForm/index.js
