+++ src/views/generalSettings/index.css
